refactor: rename `@core` to `core`
