+++ server/src/core/scalar/uuid.scalar.ts
