+++ server/src/infra/http/resolvers/list-submissions.args.ts
@@ -1,5 +1,5 @@
 import { ArgsType, Field, InputType } from '@nestjs/graphql';
-import { PaginationArgs } from '~/common/pagination/pagination.args';
+import { PaginationArgs } from '~/core/pagination/pagination.args';
 import { SubmissionStatus } from '~/domain/submissions/entities/submission.entity';
 
 @InputType()
